requirements.txt
ruff
streamlit
llama-index
llama-index-llms-openai
GitPython
streamlit-pills